fix/lag cada 5m arreglado con worker threads
