+++ package.json
@@ -2,7 +2,6 @@
   "name": "compet_app",
   "version": "0.1.0",
   "private": true,
-  "type": "module",
   "browser": {
     "fs": false,
     "child_process": false
