+++ src/mrp_data/query_mrp_data.ts
@@ -1,6 +1,4 @@
 /* eslint-disable */
-import "server-only";
-
 import dayjs from "dayjs";
 import { getSetting } from "~/lib/settings";
 import type {
