+++ src/scripts/lib/database.ts
@@ -1,4 +1,3 @@
-import "server-only";
 import sql from "mssql";
 import {
   clientSchema,
@@ -20,12 +19,12 @@ import {
   productSchema,
   productStockCommitedSchema,
   providerSchema,
-} from "~/lib/types";
+} from "../../lib/types";
 import { z } from "zod";
 import { soldProductsQuery, soldProductsQueryByCode, soldQuery } from "./large-queries";
-import { cachedAsyncFetch } from "~/lib/cache";
-import { env } from "~/env";
-import { queryBaseMRPDataUT } from "~/serverfunctions";
+import { cachedAsyncFetch } from "../../lib/cache";
+import { env } from "../../env";
+import { queryBaseMRPDataUT } from "../../serverfunctions";
 
 const defaultConnectionQuery =
   process.env.CONNECTION_QUERY ??
