+++ src/scripts/lib/instance.ts
@@ -1,16 +1,16 @@
-import { env } from "~/env";
+import { env } from "../../env";
 import { Database } from "./database";
 
 let dbInstance: Database | null = null;
 
 export const getDbInstance = async (): Promise<Database> => {
-    if (dbInstance === null) {
-        const db = new Database();
-        if (env.DB_DIRECT_CONNECTION) {
-            await db.open();
-        }
-        dbInstance = db;
+  if (dbInstance === null) {
+    const db = new Database();
+    if (env.DB_DIRECT_CONNECTION) {
+      await db.open();
     }
+    dbInstance = db;
+  }
 
-    return dbInstance;
-}
+  return dbInstance;
+};
