+++ src/scripts/load-data.ts
@@ -5,13 +5,15 @@ import { loadDataFromTangoToCloud } from "./lib/load-data-func";
 
 console.log(`[${dayjs().format("DD/MM/YYYY HH:mm:ss")}]`);
 
-try {
-  await loadDataFromTangoToCloud();
-} catch (error) {
-  console.error(error);
-  process.exit(1);
-}
+void (async () => {
+  try {
+    await loadDataFromTangoToCloud();
+  } catch (error) {
+    console.error(error);
+    process.exit(1);
+  }
 
-console.log("Saliendo...");
+  console.log("Saliendo...");
 
-process.exit(0);
+  process.exit(0);
+})();
