+++ src/worker.js
@@ -1,0 +1,4 @@
+const path = require('path');
+
+require('ts-node').register();
+require(path.resolve(__dirname, './worker.ts'));
